pgxp: compute NCLIP as a double-precision factored cross product
PGXP_NCLIP (used when the PGXP NCLIP option is enabled) computed the
back-face cross product as six separate float products summed together:
SX0*SY1 + SX1*SY2 + SX2*SY0 - SX0*SY2 - SX1*SY0 - SX2*SY1. For the thin
and near-degenerate triangles typical of PSX geometry the positive and
negative terms nearly cancel, so accumulating them in 24-bit float loses
most of the significant bits - measured error up to ~1.4 against the true
cross product, with mean ~0.06. The result is truncated to int32 and
written to GTE MAC0, so that error is fully realised (and the pre-existing
0.1<|x|<1 nudge is a band-aid for the same cancellation manufacturing
spurious near-zero sums).

Rewrite it in double using the factored form
  SX0*(SY1-SY2) + SX1*(SY2-SY0) + SX2*(SY0-SY1)
- the same algebra as the native integer NCLIP - which is accurate to the
true cross product of the (float) shadow coords (error ~0 vs reference
over 2M randomised triangles incl. near-degenerate). Return type widened
to double; the caller's int32 truncation and the small-value nudge are
unchanged in intent.

This tightens PGXP culling accuracy; it does not change the determinism
characteristics of the NCLIP-into-MAC0 feedback (still float-derived), and
is inert unless the PGXP NCLIP option is enabled.

+++ pgxp/pgxp_gte.h
@@ -48,7 +48,7 @@ void	PGXP_pushSXYZ2s(int64_t _x, int64_t _y, int64_t _z, uint32_t v);
 void	PGXP_RTPS(uint32_t _n, uint32_t _v);
 
 int		PGXP_NCLIP_valid(uint32_t sxy0, uint32_t sxy1, uint32_t sxy2);
-float	PGXP_NCLIP();
+double	PGXP_NCLIP();
 
 // Data transfer tracking
 void	PGXP_GTE_MFC2(uint32_t instr, uint32_t rtVal, uint32_t rdVal);		// copy GTE data reg to GPR reg (MFC2)
